+++ CHANGELOG.md
@@ -7,6 +7,10 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ---
 
+## 2025-01-21
+
+- Add analysis for fragpipe results
+
 ## 2025-01-20
 
 - Add CombineFasta, PhilosopherFilter, and ComputeFDR.